Fixed various issues in invoice page

+++ PeaceInternational.Core/Entity/Invoice.cs
@@ -21,9 +21,9 @@ namespace PeaceInternational.Core.Entity
         public int PAX { get; set; }
         public string Guide { get; set; }
         public string Vehicle { get; set; }
-        public double TotalDue { get; set; }
-        public double Discount { get; set; }
-        public double NetAmount { get; set; }
+        public decimal TotalDue { get; set; }
+        public decimal Discount { get; set; }
+        public decimal NetAmount { get; set; }
 
         public Customer Customer { get; set; }
